cleaned up and renamed merge script
